improving chronos runtime for append and list
chronos append and iterate are vastly sped up by using sendfile(2) and
splice(2) instead of read(2)/write(2). The speed up depends mostly on the size
of values that have to be copied out of the data file to stdout.

Assuming we use the same buffer size as what we'd use for read(2) and write(2)
then splice(2)/sendfile(2) would use half as many context switches to the
kernel. Since keeping everything in kernel space allows the kernel to use any
buffer size it deems appropriate, we can save even more time than that.