Reorder command comparisons to fast path append
I turns out that strncmp is a little slow. So we optimize the order of our
commands to ensure that we can run as fast as possible for the high performance
case (append) and take just a little bit longer on all the other cases. The
ordering is based solely on what I expect to be used more often, rather than
actually profiling a use case of chronos.