Daemon cleans up when being killed by TERM and INT
Daemon code cleans up after itself when sent the sigterm or sigint signals, it
does so by unlinking the fifo that it created earlier. This works, because the
fifo was guaranteed to have been created by the current process (so it never
starts when a fifo already exists), and can not be created by any other process
(since the current process holds the write flock(2) on the directory).
Additionally unlinking the fifo is used to signal to the open-fifo-read-close
loop to exit, since the call to open(2) will fail once the fifo has been
unlinked.

Also some tests for the daemon mode.