Adds not-pipe message to chronos-append
chronos_append takes an input file descriptor and will return an
error if it was not acceptable for use in splice(2). Previously this
error was C_IO_READ_ERROR, a generic catch all for "splice failed".

This change some checks to detect if the inputted file descriptor may
not have been acceptable (since the rest of the splice(2) arguments
are likely fine), and then prints out an error message to the user if
this is the case.